// web/src/api/models/TypesProcessInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesProcessInfo
 */
export interface TypesProcessInfo {
    /**
     * 
     * @type {string}
     * @memberof TypesProcessInfo
     */
    cmd?: string;
    /**
     * 
     * @type {number}
     * @memberof TypesProcessInfo
     */
    cpu_pct: number;
    /**
     * 
     * @type {number}
     * @memberof TypesProcessInfo
     */
    mem_rss: number;
    /**
     * 
     * @type {string}
     * @memberof TypesProcessInfo
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof TypesProcessInfo
     */
    pid: number;
    /**
     * 
     * @type {string}
     * @memberof TypesProcessInfo
     */
    state: string;
}

/**
 * Check if a given object implements the TypesProcessInfo interface.
 */
export function instanceOfTypesProcessInfo(value: object): value is TypesProcessInfo {
    if (!('cpu_pct' in value) || value['cpu_pct'] === undefined) return false;
    if (!('mem_rss' in value) || value['mem_rss'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('pid' in value) || value['pid'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    return true;
}

export function TypesProcessInfoFromJSON(json: any): TypesProcessInfo {
    return TypesProcessInfoFromJSONTyped(json, false);
}

export function TypesProcessInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesProcessInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'cmd': json['cmd'] == null ? undefined : json['cmd'],
        'cpu_pct': json['cpu_pct'],
        'mem_rss': json['mem_rss'],
        'name': json['name'],
        'pid': json['pid'],
        'state': json['state'],
    };
}

export function TypesProcessInfoToJSON(json: any): TypesProcessInfo {
    return TypesProcessInfoToJSONTyped(json, false);
}

export function TypesProcessInfoToJSONTyped(value?: TypesProcessInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cmd': value['cmd'],
        'cpu_pct': value['cpu_pct'],
        'mem_rss': value['mem_rss'],
        'name': value['name'],
        'pid': value['pid'],
        'state': value['state'],
    };
}